Split out position constrainer and dialogue activation of it
