+++ Assets/Prefabs/Collision/DialoguePoisitionConstrainerActivator.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b73a629ced8cdd84380f12fc86a7a986
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
