+++ Assets/Prefabs/Collision/PositionConstrainer.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &4561479639962204881
+--- !u!1 &4833272182958045684
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,48 +8,48 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 4561479639962204886}
-  - component: {fileID: 4561479639962204885}
-  - component: {fileID: 4561479639962204884}
-  - component: {fileID: 4561479639962204887}
-  - component: {fileID: 4561479639962204890}
+  - component: {fileID: 4833272182958045683}
+  - component: {fileID: 4833272182958045680}
+  - component: {fileID: 4833272182958045681}
+  - component: {fileID: 4833272182958045682}
+  - component: {fileID: 4833272182958045695}
   m_Layer: 0
-  m_Name: DialoguePoisitionConstrainer
+  m_Name: PositionConstrainer
   m_TagString: Untagged
   m_Icon: {fileID: 5132851093641282708, guid: 0000000000000000d000000000000000, type: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4561479639962204886
+--- !u!4 &4833272182958045683
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4561479639962204881}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 209.761, y: 380.299, z: -259.379}
+  m_GameObject: {fileID: 4833272182958045684}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!33 &4561479639962204885
+--- !u!33 &4833272182958045680
 MeshFilter:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4561479639962204881}
+  m_GameObject: {fileID: 4833272182958045684}
   m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
---- !u!23 &4561479639962204884
+--- !u!23 &4833272182958045681
 MeshRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4561479639962204881}
+  m_GameObject: {fileID: 4833272182958045684}
   m_Enabled: 0
   m_CastShadows: 1
   m_ReceiveShadows: 1
@@ -85,34 +85,29 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!65 &4561479639962204887
+--- !u!65 &4833272182958045682
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4561479639962204881}
+  m_GameObject: {fileID: 4833272182958045684}
   m_Material: {fileID: 0}
   m_IsTrigger: 1
   m_Enabled: 1
   serializedVersion: 2
   m_Size: {x: 1.9438171, y: 1, z: 1}
   m_Center: {x: -0.07872009, y: 0, z: 0}
---- !u!114 &4561479639962204890
+--- !u!114 &4833272182958045695
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4561479639962204881}
+  m_GameObject: {fileID: 4833272182958045684}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f72604004e3a05a4a8eb02b6fef17ab1, type: 3}
+  m_Script: {fileID: 11500000, guid: 4aaec1a6098060d46a742049290fe039, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  settings:
-    <ActivateId>k__BackingField: 0
-    <DeactivateId>k__BackingField: 0
-    <PlayId>k__BackingField: 1
-  active: 1
   offsetDistance: 2
