+++ Assets/Prefabs/Collision/PositionConstrainer.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1185f8ad9fa8bb6468027bf88a8f1920
+guid: b3484ce6a5d9ac346bac9b1f584d222b
 PrefabImporter:
   externalObjects: {}
   userData:
