+++ Assets/Scripts/Environment/PositionConstrainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4aaec1a6098060d46a742049290fe039
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
